[RtCA][22/11/2015] Version 0.6.899.913b Adding functionalities:   - Review few bugs   - Review bug in ShellBags file extract   + Add tag on session name (terminal and GUI)   + Backup files from list (local search files, imported files or from a system root)
Temporary disable:
- no x64 binary update

+++ RTCA/tools/tools_virustotal.c
@@ -404,7 +404,7 @@ DWORD WINAPI CheckAllFileToVirusTotalProcess(LPVOID lParam)
   hSemaphoreItem=CreateSemaphore(NULL,1,1,NULL);
 
   //la gestion du nombre en lecture continue permet d'effectuer un scan pendant l'énumération
-  DWORD i;
+  long int i;
   vt_error=0;
 	sv.token[0] = 0;
   for (i=0;i<SendMessage(hlstv_process,LVM_GETITEMCOUNT,(WPARAM)0,(LPARAM)0);i++)
